fix: type cast back to float64 for date additions.

+++ variant.go
@@ -667,7 +667,7 @@ func TimeToVariant(i any) *VARIANT {
 	seconds := duration.Seconds()
 	milliseconds := duration.Milliseconds()
 
-	winDate := float64(days) + float64(hours*DateSingleHour) + float64(minutes*DateSingleMinute) + float64(seconds*DateSingleSecond) + float64(milliseconds*DateSingleMilliSecond)
+	winDate := float64(days) + float64(hours*float64(DateSingleHour)) + float64(minutes*float64(DateSingleMinute)) + float64(seconds*float64(DateSingleSecond)) + float64(milliseconds*float64(DateSingleMilliSecond))
 	number := math.Float64bits(winDate)
 
 	return &VARIANT{VT: VT_DATE, Val: int64(uintptr(unsafe.Pointer(number)))}
